Refactor to use kubernetes cronjobs
Recommended practice for running cronjobs in Kubernetes
https://kubernetes.io/docs/concepts/workloads/controllers/cron-jobs/

Why run cron as CronJob resource in kubernetes?

Scaling

Sidecar is not a good practice for crons as it is difficult to support multiple replicas.
I.e. if run with x replicas it means the same cron would run in paralell x times.
Cronjobs in kubernetes support concurrencyPolicy: Forbid to solve this issue.

Monitoring

Logging works out of the box and successful/failed jobs can be monitored

Resources

Using cronjobs allow better flexibility in case of resources in the cluster.
For instance, jobs can run on dedicated nodes (nodeSelector), custom resources etc

Signed-off-by: Mikael Hammarin <mhammarin@ecit.com>